Crate sgx_tcrypto

Cryptography Library

The Intel(R) Software Guard Extensions SDK includes a trusted cryptography library named sgx_tcrypto. It includes the cryptographic functions used by other trusted libraries included in the SDK

Structs

SgxAesHandle

SgxCmacHandle
CMAC algorithm context state.

SgxEccHandle
ECC GF(p) context state.

SgxHmacHandle

SgxRsaPrivKey

SgxRsaPubKey

SgxSha1Handle

SgxShaHandle
SHA algorithm context state.

Constants

SGX_AESCTR_CTR_SIZE

Functions

rsgx_aes_ctr_decrypt
rsgx_aes_ctr_decrypt performs a Rijndael AES-CTR decryption operation.

rsgx_aes_ctr_encrypt
rsgx_aes_ctr_encrypt performs a Rijndael AES-CTR encryption operation.

rsgx_align_ecc256_priv_key

rsgx_align_hmac_sha256_msg

rsgx_align_hmac_sha256_slice

rsgx_calculate_ecdsa_priv_key
rsgx_calculate_ecdsa_priv_key generates an ECDSA private key based on an input random seed.

rsgx_create_rsa_key_pair

rsgx_ecc256_calculate_pub_from_priv

rsgx_ecc256_priv_key

rsgx_ecc256_pub_from_priv

rsgx_hmac_sha256_msg

rsgx_hmac_sha256_slice

rsgx_rijndael128GCM_encrypt
rsgx_rijndael128GCM_encrypt performs a Rijndael AES-GCM encryption operation.

rsgx_rijndael128GCM_decrypt
rsgx_rijndael128GCM_decrypt performs a Rijndael AES-GCM decryption operation.

rsgx_rijndael128_cmac_msg
The rsgx_rijndael128_cmac_msg function performs a standard 128bit CMAC hash over the input data buffer.

rsgx_rijndael128_align_cmac_msg

rsgx_rijndael128_cmac_slice
The rsgx_rijndael128_cmac_slice function performs a standard 128bit CMAC hash over the input data buffer.

rsgx_rijndael128_align_cmac_slice

rsgx_rsa3072_sign_msg
The rsgx_rsa3072_sign_msg computes a digital signature for a given dataset based on RSA 3072 private key.

rsgx_rsa3072_sign_slice
The rsgx_rsa3072_sign_slice computes a digital signature for a given dataset based on RSA 3072 private key.

rsgx_rsa3072_verify_msg
rsgx_rsa3072_verify_msg verifies the input digital signature for the given data- set based on the RSA 3072 public key.

rsgx_rsa3072_verify_slice
rsgx_rsa3072_verify_slice verifies the input digital signature for the given data- set based on the RSA 3072 public key.

rsgx_sha1_msg

rsgx_sha1_slice

rsgx_sha256_msg
The rsgx_sha256_msg function performs a standard SHA256 hash over the input data buffer.

rsgx_sha256_slice
The rsgx_sha256_slice function performs a standard SHA256 hash over the input data buffer.

Type Definitions

sgx_aes_ctr_128bit_ctr_t