Function sgx_tcrypto::rsgx_calculate_ecdsa_priv_key

pub fn rsgx_calculate_ecdsa_priv_key(
    hash_drg: &[u8], 
    sgx_nistp256_r_m1: &[u8], 
    out_key: &mut [u8]
) -> SgxError

rsgx_calculate_ecdsa_priv_key generates an ECDSA private key based on an input random seed.

Description

This function generates an ECDSA private key based on an input random seed.

Parameters

hash_drg

Pointer to the input random seed.

sgx_nistp256_r_m1

Pointer to the buffer for n-1 where n is order of the ECC group used.

out_key

Pointer to the generated ECDSA private key.

Requirements

Library: libsgx_tcrypto.a

Errors

SGX_ERROR_INVALID_PARAMETER

Some of the pointers are NULL, or the input size is 0.

SGX_ERROR_UNEXPECTED

Unexpected error occurred during the ECDSA private key generation.